Add unit tests for skill registry and symbol validation
- Introduced `test_registry.py` to validate the on-disk skill format and discovery layer.
- Added tests for bundled skills, ensuring they are discovered correctly and contain valid metadata.
- Implemented validation tests for skill frontmatter, including checks for missing or malformed keys.
- Created `test_symbols.py` to verify that all symbols mentioned in skills exist in the railtracks package.
- Enhanced `test_cli.py` by removing deprecated tests and ensuring the CLI functions correctly with skills.
