+++ packages/railtracks/tests/unit_tests/cli/skillkit/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for `railtracks.cli._skillkit` — the skill format, manifest and installer."""
